requirements.txt
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
decorator==5.1.1
Developing-Cockroach==0.4.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
h11==0.14.0
idna==3.4
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
openpyxl==3.0.10
outcome==1.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PySocks==1.7.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wcwidth==0.2.6
wsproto==1.2.0

